Fix unsigned 32-bit offset overflow in zCom4/zText4/RawCom4 drivers (#157)
The *4 drivers (zCom4, zText4, RawCom4, RawLD4) use 4-byte unsigned
integers for file offsets. Java's int is signed, so values above 2GB
appear negative, causing RandomAccessFile.seek() to fail on large modules.

Changes:
- DataIndex: change offset field/constructor/getter from int to long
- ZVerseBackend: use & 0xFFFFFFFFL on blockNum (prevents sign-extension
  int->long) and change blockStart from int to long
- RawBackend.getIndex(): change entryOffset from int to long with
  unsigned masking
- RawLDBackend.getIndex(): same fix; update debug dump methods to long
- ZLDBackend.getEntry(): change blockStart from int to long with
  unsigned masking; update debug dump method to long
- DataEntry.getBlockIndex(): apply & 0xFFFFFFFFL to block offset when
  constructing DataIndex
- gradle-wrapper: upgrade to Gradle 8.10 for Java 23+ compatibility

+++ src/main/java/org/crosswire/jsword/book/sword/DataIndex.java
@@ -35,7 +35,7 @@ public class DataIndex {
      * @param size
      *            The number of bytes to read from the file.
      */
-    public DataIndex(int offset, int size) {
+    public DataIndex(long offset, int size) {
         this.offset = offset;
         this.size = size;
     }
@@ -43,7 +43,7 @@ public class DataIndex {
     /**
      * @return the offset
      */
-    public int getOffset() {
+    public long getOffset() {
         return offset;
     }
 
@@ -54,6 +54,6 @@ public class DataIndex {
         return size;
     }
 
-    private int offset;
+    private long offset;
     private int size;
 }
